:recycle: CONTRAST-24200: statically link libprotobuf-c
Statically link libprotobuf-c with the nginx connector dso so we do not
rely on the system's package repo for it and we won't conflict with the
system package version if its already there.
Add libprotobuf-c v1.3.0 tag as a git submodule.
Make build use the protobuf-c submodule rather than the bootstrap
version.
Install the protobuf-c lib in a build/ dir so that it doesn't require
sudo privs and keeps the system clean.
Remove protobuf-c from bootstrap script. git-submodule will manage it
now.
Add submodule init and other helpers to make compilation a little easier
with the protobuf-c changes.