:package: CONTRAST-24228: create ubuntu packaging scripts
This changeset will add debian package creation to the nginx connector
along with processes to help make debian packages.  While working on
making the package building work, I reduced dependencies and added
documentation.

* Fix compiler warnings so that debian build systems doesn't error out on
it.
* Add usage of the oss-pkg tool from NGINX with modified NGINX supplied
build_module.sh script.
* Add Notes file to explain and clarify what is not explained or clarified
in all of the NGINX provided tools/scripts.  Also discusses future
development direction.
* Add Makefile to build the contrast module deps (protobuf-c) for easier
integration with the deb packaging helpers.
* Add more ubuntu revisions to Vagrant file (14.04 - 18.04) for package
building and testing. There is an 'ubuntu-test' vagrant image which keeps
hardcoded external port mappings for easier testing.
* Change protobuf-c dependency to not build the protoc-c compiler which
Removed the dependency on google libprotoc >= 3.0.0 altogether. We only
need the static library of it.
* Change module conf to know where to find the internally built protobuf-c
headers so it can correctly compile the pre-generated protobuf files.
Note this may need to change to be more dynamic in the future and
generate the protobuf C files at build time.
* Remove protobuf from bootstrapping env since this modules doesn't
officially depend on it.
* Make local build use local headers from local protobuf-c build.