Fix package versioning to target different nginx version.
The project previously used the nginx version as its own version. This
was problematic because we couldn't effectively version the module
independently of nginx and release updates for the module in a sane way.

This changeset incorporate nginx and module version into a single package
version. Furthuremore, the version of this module is dynamically
calculated so there is no need to constantly bump a VERSION file. We
attempt to version according to semver.org. The dynamic version creation
uses the git tag to create a MAJOR.MINOR.PATCH version number. If the
version is being calculated on a non-tagged commit, the algorithm of
version creation will create
    <parent tag version>+<commit num since tag>dev.<commit hash>

So if version 1.0.4 was just tagged, the version of that commit would be
1.0.4.

If I begin development again on master, and have since made two
additional commits, the version number would be:
    1.0.4+2dev.gac76cb3e

Versioning in this manner will allow our internal developmental packages
to override the previous release package. When a new tag is made on a
hypothentical developmental branch as above, it will be 1.0.5, which
overrides all previous developmental branches.

The commit hash is used in developmental versioning because multiple
branches can split from a single parent tag and the commit number
distance from the parent is not unique in this case.  Hence a commit
hash is added to make it unique. Really the other numbers are just for
easy human consumption.

When building this module, a specific version of nginx must be targeted.
So to allow a single release of the module to target multiple versions
of nginx, I prefix the module version number with the targeted nginx
version.  This is a little ugly, but will allow a human to understand
versioning of a package and install the appropriate version for thier
system should they choose an nginx older than the latest stable release.

If I build this module targeting nginx 1.14.0, then the final package
version using our hypothetical module version above is:
    1.14.0+1.0.4+2dev.gac76cb3e

and a release version would look like:
    1.14.0+1.0.5

Note, version number construction avoids the use of '_', '-', and '~' to
remain compatible with both debian and rpm versioning guidelines.

related tickets:
    CONTRAST-23167