add support to pkg-oss tool for non-nginx-standard naming.
The oss-pkg tool from nginx forces the module naming pattern of
nginx-module-<module_name>. To be consistent with the other package
names from Contrast, we prefer this name to be:
    contrast-webserver-agent-nginx

This changeset will add support to the nginx oss-pkg tool to allow a
user to specify the pkg the name directly, and then it makes use of it
to force the new package naming convention.

After all this work, in hindsight, it probably would have been easier to
ignore the oss-pkg tool and just perform packaging directly ourselves.
But oh well, hindsight is 20/20.

Removed legacy version 1.12.2 from pipeline. We are targeting the
lastest stable distribution from nginx directly currently

Related ticket:
    CONTRAST-23167