Add support for file upload analysis in the nginx connector.
Previously, 'expect: 100-continue' transaction were not being properly
buffered before sending the request over to speedracer, so speedracer
would only see a partial malformed request and fail it.  File uploads
are put into an http body as a multipart/form-data with each 'part'
having its own content-type.

This changeset now properly buffers the entire 'expect: 100-continue'
transaction before sending to speedracer so speedracer will process the
entire request, which includes the file upload. There is a bug in the
underlying protobuf used to transport the http request body in that it
cannot handle binary data, which would be the case on a content-type
application/octet-stream. This happens if you try to upload a binary
file.  This bug will be fixed in a future changeset.

This changeset will also enable http2 in nginx server builds. It is
turned on in the config if you want it. This change just enables the
supporting code in nginx.

Related Tickets:
    CONTRAST-23461