Allow binary http content in speedracer comms
Prior to this change, a binary file uploaded could not be sent to
speedracer for analysis because the underlying protobuf used to send the
data only supported null-terminated string values for the body.

This change makes the http request/response body fields in protobuf
comms to speedracer be binary/bytes types rather than string types. This
supports the binary content type application/octet-stream and allows for
uploaded files to be sent to speedracer for analysis.

Also updated new location of socket file in test config

Related Ticket:
    CONTRAST-26609