Add 'make dist' and build from src capability.
Prior to this changeset our build process was very tightly coupled to
working from a git repo and installing nginx as we do for testing and
development. We want to offer this product as a build-from-source
capability so consumers can choose to build a statically compiled module
into nginx or build nginx with different options than we do for test and
development.

The changeset separates out our external dependency on protobuf-c-1.3.1
and organizes things so that our module build follows common nginx
module build instructions. This allows our module build mechanics to be
familiar to people who work with building nginx with custom modules.
Also the method of how a remote external library (protobuf-c) is
included as part of a configure script is more similar.

Furthermore, the 'make dist' target is added so a developer can create a
source distribution of the connector module. This tarball leaves out all
the developer helper scripts and what not and only includes the source
necessary for module building in the nginx build system.

One other thing is that the change also formally adds the ability to
build a static module into nginx. Prior to this we just developed a
dynamic module.

Related Tickets:
    CONTRAST-27190