Apply sane defaults for connector
The connector module currently wants about 4 additional config lines
after a user has added the load_module directive. To make this easy for
the first time user and aid in the TS wizard steps, we decided to make
only the load_module directive required and the rest of the config lines
will just have reasonable defaults that will "get things working".

This changeset makes only the load_module directive required and uses
working defaults for the rest of the contrast_* nginx conf directives.
During this work, logging in a few areas were also deficient so I fixed
them and added it to this changeset. Also some of the reported errors
during socket connection were also lacking detail so I enhanced the log
msg to add the interpreted errno detail.

"just because I was there"... I also converted tabs to spaces in
*_socket.c.

Related Tickets:
    CONTRAST-28136
    CONTRAST-28086