Reduce logging noise of module
The connector module alway had debug logging, dd(), enabled and some of
the log messages were being put out at too high of a level. This would
cause a lot of logging noise in an nginx access/error.log which could
annoy a sys admin.

This changeset makes it so debug logging is only compiled and enabeld in
debug-builds of nginx. That is, builds with the '--with-debug' nginx
config option. Also some of the other contrast messages related to
blocking a request have been removed or reduced to the INFO level. We
don't want to overload an error log with attack data when that
infomation is already tracked by TeamServer.

Related Tickets:
    CONTRAST-28143