Fix operation for nginx certifcation tests
Nginx certification tests are typically performed with the module loaded
but disabled. It's verifying a loaded module doesn't mess up operation
of the nginx server as a whole. We pass the tests in this instance.

The tests also have value to ensure we don't mess up nginx operation
when the module is actually enabled and tapping the data streams.  In
this case we don't want to inadvertantly break some nginx operation in a
complex nginx config. The changeset here is fixing the module so it
passes nginx tests when its enabled.

---
Fix header filter processing

The contrast module, when on, would previously short circuit additional
outgoing header processing. This was found in the test
nginx-tests/body_chunked.t. The test isn't looking for this issue, but
uses the add_header config directive to perform the tests. Since we
short-circuited the header-filter, the dynamic 'add_header' wasn't being
done and the test failed.

This changeset fixes the short-circuit by passing control to the next
module in the chain.

---

Fix request headers were being double sent in empty_gif.t

---

mirror_proxy.t exposed issue when processing subrequests in pre-access
handler. Subrequest do not want inherited body from main request unless
r->preserve_body is set.

---

Related Tickets:
    CONTRAST-28338