Add option for response body processing
Prior to this change, our nginx module did always did response body
processing.  What this means is that it attempted to buffer the entire
response body in an nginx body filter so it could be analyszed in full
and either sent out or not depending on intervention. The nginx
certification tests didn't like this behavior, nor is it good for
performance, nor does the nginx framework like it.  But it does
functionally work.

This change adds a config option so that a user can decide whether or
not they want this behavior. By default we will not do it.  This means
that data-leakage rules will not work.

Related Tickets:
    CONTRAST-28335