fix nginx crash when module disabled
Prior to this change, the server would crash in the LOG phase if the
module was loaded but disabled. This is because the log phase was
referencing structures it should not have been when the module is
disabled in the earlier phases.

This change will make an 'enable' check in the LOG phase before doing
any work.

Related Tickets:
    CONTRAST-28421