Up default build against nginx 1.14.1
Prior to this change, all default building and pkg build would produce
binaries dependent on nginx 1.14.0.

This change makes the default produced binaries dependent on nginx
1.14.1

Reason to up the version number of nginx is because there was a CVE (DoS
I think) found in the httpv2 module of nginx 1.14.0.